# Intercept SIGINT signal on tests suites