If the `pcntl_signal()` function exists, creates a signal handler for `SIGINT` which prints the summary of test suite and generates code coverage reports.
Move all code which prints results and generates coverage from `doRun()` method to `doReports()` method on `PHPUnit_TextUI_TestRunner` class.

Here some benchmark of PHPUnit's test suite, without [tests/TextUI/sigint.phpt](https://github.com/henriquemoody/phpunit/blob/sigint-intercept/tests/TextUI/sigint.phpt) file:

**Without SIGINT interception**

PHP Version | Time          | Memory
------------|---------------|--------
5.3.3       | 28.6 seconds  | 31.50Mb
5.3         | 17.85 seconds | 31.50Mb
5.4         | 18.63 seconds | 23.00Mb
5.5         | 19.14 seconds | 23.25Mb
5.6         | 19.21 seconds | 23.50Mb

**With SIGINT interception**

PHP Version | Time          | Memory
------------|---------------|--------
5.3.3       | 20.04 seconds | 31.75Mb
5.3         | 21.18 seconds | 31.75Mb
5.4         | 21.76 seconds | 23.25Mb
5.5         | 18.62 seconds | 23.50Mb
5.6         | 20.23 seconds | 23.50Mb

**See it working**: https://asciinema.org/a/14111