@henriquemoody 2015-02-06
@sebastianbergmann, it has been a while since the last comment on this pull request.

Looks like 4.6 has been released so this change will only be available (if accepted) on 6.7 now...

@derickr said there is no problem with ticks declaration and I have already explained about `pcntl_signal_dispatch()`, also I have made benchmarks (they are in the description now) often I rebase it to be up to merge so what is missing here?