@sebastianbergmann 2015-02-06
PHPUnit 4.6 was not released but you are right that if this were merged now then it would go into PHPUnit 4.7. However, I do not really want to merge this. The negative implication of activating ticks on performance outweighs the benefit IMHO.